Fix `t.rb` working in sub-directories of the git repo
This fixes `t.rb` working in sub-directories of the git repo by
adding the `--relative` flag to the `git diff` command. For
example, if you have the following folder structure:

    gitrepo
     - .git
     - subdirectory
        - .t.yml
        - tests

and you are in the `subdirectory` folder, if you ran `t -c` to
detect changes in your `tests` folder, it previously wouldn't see
any. This is because the `git diff` command returned results like
`gitrepo/subdirectory/tests/something` and then the test running
logic wouldn't think it's a valid file for it to run. The
`--relative` option makes it return `subdirectory/tests/something`
which works as expected with the rest of the test running logic.